requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.0.1
WTForms==3.2.1
pytest==8.3.5
pytest-flask==1.3.0
